Remove pkg/serverconfig, merge to apis. Fix some deprecated builder annotations
